+++ packages/methods/src/focus/focus.ts
@@ -16,13 +16,6 @@ export interface FocusFieldConfig<
   TFieldPath extends RequiredPath,
 > {
   /**
-   * The form store containing the field.
-   *
-   * TODO: This should be the first parameter, not part of the config.
-   */
-  readonly form: BaseFormStore<TSchema>;
-
-  /**
    * The path to the field to focus.
    */
   readonly path: ValidPath<v.InferInput<TSchema>, TFieldPath>;
@@ -33,10 +26,12 @@ export interface FocusFieldConfig<
  * programmatically setting focus to a specific field, such as after
  * validation errors or user interactions.
  *
+ * @param form The form store containing the field.
  * @param config The focus field configuration.
  */
 export function focus<TSchema extends Schema, TFieldPath extends RequiredPath>(
+  form: BaseFormStore<TSchema>,
   config: FocusFieldConfig<TSchema, TFieldPath>
 ): void {
-  getFieldStore(config.form[INTERNAL], config.path).elements[0]?.focus();
+  getFieldStore(form[INTERNAL], config.path).elements[0]?.focus();
 }