+++ packages/methods/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix `focus` method to accept `form` as first parameter
+
 ## v0.3.0 (September 22, 2025)
 
 - Change `@formisch/core` to v0.4.0